Add use_syslog=True to sysinv.conf
When bootstrapping a system, the sysinv.conf does not have the
use_syslog option set in the sysinv.conf. Because of that,
sysinv.log and sysinv-api.log files are not created until the
host unlock, when the configuration file is overridden and the
option is set.
This commit adds the use_syslog=True option in the sysinv.conf
template used to bootstrap the system.

Test plan:
1. PASS: Bootstrap a DC system with three subclouds.
2. PASS: During the bootstrap process, verify that the
sysinv.log file is created after sysinv.conf is generated.
3. PASS: During the bootstrap process, verify that the
sysinv-api.log file is created after the "Saving config
in sysinv database" step.

Closes-Bug: 2068767

Change-Id: I7ce56aa56114fb3e5234913bfe8a081bc6851612
Signed-off-by: Raphael Lima <Raphael.Lima@windriver.com>