Change oidc-auth-apps-certificate secret observer
This review will be changing the secret observer of oidc-auth-apps
to look for the ca.crt value inside of oidc-auth-apps-certificate,
removing then the step to create a new secret that stores this
value.

Test Plan:
PASS: Generate a new .iso with the changes.
PASS: Run a full deploy of an AIO-SX and DC system with the
      generated iso.
PASS: Use oidc to switch to another user context and execute
      kubectl commands in AIO-SX and DC system.
PASS: Run update-platform-certificate playbook in AIO-SX and
      DC system.
PASS: Create a test job out of the daily-scheduled secret
      observer and verify the logs to see if finished
      correctly.
PASS: Delete the secret and verify if the observer will restart
      deployment in the correct timeset.

Closes-Bug: 2069323

Change-Id: Ic590fb492caf7e5bbc15851db0a9fffead7cd69b
Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>