Fix reading DC role from sysinv during enrollment
This commit fixes the ansible playbook reading the DC role from
sysinv during enrollment. During enrollment, the subcloud sysinv
is not configured as a subcloud, which results in ansible running
as a non subcloud. Also fixes other places in the playbook that
was not running as intended for enrollment.

Test Cases:

The test environment is a standard system controller with a single
node subcloud, installed with the existing factory install procedure.
The enrollment playbook was ran and the certificates checked against
an existing DC deployment. Specifically,
- dc-adminep-certificate
- dc-adminep-root-ca-certificate
- sc-adminep-certificate
- sc-intermediate-ca-adminep-issuer
- dc-adminep-root-ca-issuer
- dc-selfsigning-issuer
- system-local-ca
ensuring they are READY in cert-manager, and the common name and SANs
are configured like the normal DC deployment.

Story: 2011100
Task: 50226

Change-Id: Ic0da0314abfe578f1b383b202570a13dfa5f4d05
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>