Fix failure when mode is not set
A recent submission for subcloud enrollment added a few checks for
mode == enroll to handle special cases for the enrollment playbook.
This causes failures when mode is not defined in other execution
paths. This commit adds additional checks to see if the mode
variable is defined.

Test Cases:

Run Ansible through the enrollment playbook with both the "mode"
set, and unset. Ensure the code executes appropriately.

Story: 2011100
Task: 50226

Change-Id: Icf64b20d6e175bcab09aa3b94fea8e8c0f033435
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>