hc_vault: use always instead of rescue in prepare-env
It is observed that when "Transfer backup tarball to {{
target_backup_dir }} on controller-0" fails, the play continues.  The
observation is made when ansible does not have permission to write to
the target_backup_dir.

Rescue implies that the play can continue, and a subsequent failure
occurs.  Use 'always' instead of 'rescue' to print a message from the
validation script.

Since the always printed message is specific to the validation script,
isolate the always condition to that task.

Take the opportunity to remove "controller-0" from the failed task name;
although it's probably always going be controller-0, it could be
confusing when it is not. 'target' is sufficient because ansible logs the
target's name.

Test Plan:
PASS  ansible-lint (changes only)
PASS  the play stops when transfer fails
PASS  script is logged and the play continues when validation passes
PASS  script is logged and the play stops when validation fails

Story: 2011073
Task: 49842

Change-Id: I1e9ce7ee37096d0c7fa908565f7515690004e4cb
Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>