Avoid rechecking system-local-ca overrides
Overrides for system-local-ca used to be verified by a role with an
optional to install the RCA cert as trusted. This role was used in
some playbooks twice (one for verifying the certs and fail early and
later to install the RCA cert), which caused the repetition of the
verification tasks during the RCA cert install.

This reviews splits this task in two (one for verification and other
to install the RCA cert as trusted), avoiding repeating several tasks.

Test plan:
PASS: Deploy DC + SX subcloud w/ production config for system-local-ca
      (providing external certs).
PASS: Run playbook to update platform certificates.
PASS: Upgrade AIO-SX.

Story: 2009811
Task: 50384

Change-Id: Ib4097c66b3400513f6f211b5b1be29ad68780b04
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>