Support using ECC keys
This review will be adding support for ECC certificates in the local
issuer, by detecting the type of certificate and changing the method
of verification if the certificate/key pair match accordingly with the
type of certificate, if it is RSA or ECC.

This new check will be done by the addition of the new file,
verify-certificate-and-key-pair.yml, that concentrate all the tasks to
the verification.

Note that the RSA verification is done by CLI commands because the
current version of community.crypto used by starlingx don't have support
to extract the modulus from the certificate public key.

Test Plan:
PASS: Run build-pkgs -c -p playbookconfig.
PASS: Run build-image.
PASS: Run a full deploy of an AIO-SX with a ECC certificate override.
PASS: Run a full deploy of an AIO-SX with a RSA certificate override.
PASS: Run a full deploy of an AIO-SX with no certificate override.
PASS: Push and pull images from docker registry.
PASS: Use system commands to list images from docker registry.
PASS: Run ldapusersetup.

Story: 2009811
Task: 50538

Change-Id: I81c7959d43eb592cf93d7efec6ed10f66f939e8d
Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>