Create dcmanager user during enrollment
In the case of subcloud enrollment, the dcmanager user is not created
in Keystone because the subcloud is not considered a subcloud until
the sysinv databse is updated. This commit makes the enrollment
playbook create the dcmanager user.

Hieradata is populated with the password for dcmanager regardless
of the system type. We use the password stored in hieradata and
keyring to ensure the user created uses the same credentials as if
the system was bootstrapped as a subcloud.

Test Cases:

Run Ansible through the enrollment playbook, ensure dcmanager user
is created.

Run the enrollment playbook again, to ensure the existing
dcmanager user does not break the playbook.

Create an openrc file for dcmanager using credentials stored in
keyring. Ensure the credentials are correct, and dcmanager can
run commands (system host-list)

Ensure dcmanager has admin roles in admin and services projects,
like a regular subcloud.

Story: 2011100
Task: 50226

Change-Id: Ie3e15abec44225f65d70b2e10a078c185ea0d14f
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>