Use Internal Endpoint for CGTSClient Session
When configuring a subcloud, the Ansible playbook requires the primary
management IP address to complete the configuration.
The script get_network_addresses_from_sysinv.py was created for this
purpose; it establishes a CGTS session to connect to Sysinv and
retrieve the primary management IP address.
The script uses the admin endpoint for the connection.

This setup works fine when the subcloud is configured using the
management network. However, during subcloud configuration using the
admin network, the admin network IP has not yet been configured in the
system, causing the admin endpoint to be set to a nonexistent IP.
This prevents the CGTS session from being established,
resulting in an error in the Ansible playbook.

Since the get_network_addresses_from_sysinv.py script always runs on
the target node, it can utilize the internal endpoint, which
consistently uses the management IP.
This management IP is already configured in the system.

To avoid similar issues in the future, the admin network IP will be set
on the loopback interface during the bootstrap process, similar to the
management and cluster networks. After the system unlocks, the admin
network will be assigned to the correct interface.

Test Done:

DC with AIO-SX subcloud fresh install using mgmt network
DC with AIO-SX subcloud fresh install using admin network

Story: 2011027
Task: 50474

Change-Id: Ic84521413a6748101998b0fefae15fbe16ae598f
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>