Add admin network support to enroll playbook
This commit introduces a new method to create the
admin network and system controller networks when
running the enroll playbook as a standalone system.
And updates the OpenStack endpoints on this system
after updating the admin networks.

Test Plan:

PASS: Verify that admin network and system controller
networks are created successfully in the system. Verified
the region config is set to True. And the admin endpoints
are updated

Story: 201100
Task: 50477

Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>
Co-signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>

Change-Id: I9163ebba35b3262197892b72d8b7ba8bcedb017f