Prevent empty commits from being taken from the metadata
As part of subcloud prestage based on USM feature, an
improvement is added to prevent commits that are empty from
being considered.

The hash of the commits are taken from the metadata, but the
base tag exists, whose commit is empty, causing an issue
when comparing the commits.

Test Cases
PASS
 - Upload a patch using software upload command
 - Create a subcloud prestage passing --for-sw-deploy param

Story: 2010676
Task: 50510

Change-Id: I622d6d197a062e3100bc1c7b2f4e8a11b17e05c3
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>