Add replay support to bringup IPsec task
ansible bootstrap playbook supports replay even when the previous
play is successful. This change added replay support to IPsec bringup
task so that it won't fail replay in the case where the previous
play is successful.

Test Plan:
PASS: Run ansible playbook to bootstrap the first controller of a DX
      system, verify the bootstrap is successful.
PASS: Re-run ansible bootstrap playbook in the system, verfiy the
      second bootstrap is also successful, and the IPsec configuration
      is generated properly.

Story: 2010940
Task: 50519

Change-Id: Ibfafcc2280c3417e1f877dfdb5850cd824693cdd
Signed-off-by: Andy Ning <andy.ning@windriver.com>