Fix IPSec configuration during B&R
When IPSec is configured in a node, a flag is created in
sysinv database. So, when we make a backup, these flags
are copied and we need to clear the flags related to all
nodes, except controller-0, because, when the database is
restored, the IPSec flag wil be restored do "Enabled",
which will fail IPSec configuration when the restored host
reboot.

There was an script to clear mgmt_ipsec inside bootstrap
role and it was relocated to a common place to be used in
both situations with a few enhancements.

Test Plan:
PASS: Full build, system install, bootstrap and unlock DX system w/
      unlocked enabled available status. Create a backup file and
      then restore it following the backup procedure. Observe that
      IPSec is enabled and SAs are established between the nodes.
PASS: Run ansible playbook to bootstrap the first controller of a DX
      system, verify the bootstrap is successful.
PASS: Re-run ansible bootstrap playbook in the system, verfiy the
      second bootstrap is also successful, and the IPsec configuration
      is generated properly.

Story: 2010940
Task: 50539

Change-Id: Ia83a04f48c6b3992890e9f15965b39dccf919c96
Signed-off-by: Leonardo Mendes <Leonardo.MendesSantana@windriver.com>