Add user provided system-local-ca flag
Include extra flags in the user config folder, signaling that the
data for system-local-ca was provided by user (in bootstrap or via the
update platform certificates playbook).

With this, we can verify in the health check for a future upgrade if
the user has provided the recommended data for system-local-ca.

Test plan:
PASS: Deploy AIO-SX w/ lab config for the platform certificates.
      Verified that the upgrade health check fails.
PASS: Deploy AIO-SX w/ production config for the platform certificates.
      Verified that the upgrade health check passes.
PASS: Upgraded AIO-SX from stx 9 to stx 10. Verified that the upgrade
      health check passes (upgraded systems are expected to be
      migrated).

Story: 2009811
Task: 50540

Change-Id: I982b2e7826ced624e7e3a5a7f1288093f48f6e28
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>