Update keystone data during subcloud enrollment
This commit introduced tasks to migrate the keystone data during
subcloud enrollment:
1. Creates the keystone user(dcmanager) that doesn't existed in a
non-subcloud controller.
2. Recovers sysinv services before the other operations as they are
required for the following tasks.
3. Removes the pre-checks of the management affecting alarms as it is
expected the subcloud will have alarms caused by a previous failed
enrollment or the loss of standby controllers as expected.
4. Checks the keystone data and avoid to update them again if the
migration was done by a previous failed operation.

Test plan:
1. Build a stx iso with this change.
2. Deploy two DCs with the new iso.
3. Build prestaged iso and seed iso based on the stx-iso, and bring up
an AIO-SX with the images.
4. Rehome a subcloud between the DC, verified the subcloud rehomed
successfully, verified the subcloud can be online, and the endpoints
can be in-sync(not including the usm, the migration of the usm is
expected to be integrated in:
https://review.opendev.org/c/starlingx/ansible-playbooks/+/923503).
5. Enroll the AIOSX, verified the playbook successfully finished.
Verified the keystone passwords are updated in both keystone and
keyring, verified the services are still active, verified the dcmanager
user is created.
6. Manually change the sysinv password in /etc/sysinv/sysinv.conf to
fail the sysinv services, verified the enroll playbook can recover the
services and finished successfully.
7. Re-run the enroll playbook, verified the keystone passwords update
are skipped as the passwords are already updated.

Story: 2011100
Task: 50551

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: Ia2c6feb9e1408f089167e9bdd56e7b11802f5782