Update metadata location for 'for-install' use case
Prior to this commit, patches metadata was copied
to /opt/software directory on the subcloud for
both "--for-install" and "--for-sw-deploy" cases
which is incorrect.

When a subcloud is prestaged for install, the patch
metadata if exist should be copied from the source
directory to /opt/platform-backup/<sw-version>/
patches/metadata.

If the user-specified prestage software version is
not the same as the subcloud software version,
the metadata are copied from the system controller
to the subcloud. Otherwise, the metadata are copied
from the local source directory to /opt/platform-backup.
For USM, the metadata are additionally copied from the
system controller to the subcloud to ensure it has the
latest minor release metadata.

The metadata source directory is /opt/patching in the
past releases and is /opt/software current and future
releases (USM-enabled).

This commit also removes CentOS related code from the
playbook as this code is longer relevant.

Test-Plan:

PASS:
1) Bring up DC-system(24.09) with one subcloud(24.09).
2) Apply the USM patch on the system controller.
3) Prestage a subcloud 'for-install' use-case.
4) Prestage successfull.
5) Verify ostree_repo is in /opt/platform-backup/
   sw-version/ directory.
6) Verify metadata is in /opt/platform-backup/
   sw-version/patches/metadata directory.

PASS:
1) Bring up DC-system(24.09) with one subcloud(24.09).
2) Apply the USM patch on the system controller.
3) Prestage a subcloud 'for-sw-deploy' use-case.
4) Prestage successfull.
5) Verify ostree_repo is in /opt/platform-backup/
   sw-version/ directory.
6) Verify metadata is in /opt/platform-backup/
   sw-version/patches/metadata directory.

PASS:
1) Bring up DC-system(24.09) with one subcloud(22.12).
2) Apply the USM patch on the system controller.
3) Prestage a subcloud 'for-install' use-case.
4) Prestage successfull.
5) Verify ostree_repo is in /opt/platform-backup/
   sw-version/ directory.
6) Verify metadata is in /opt/platform-backup/
   sw-version/patches/metadata directory.

Closes-Bug: 2075312

Change-Id: Ia68280096ef4d4cefbad98676d570ac0e579b81e
Signed-off-by: sbhardwa <sharang.bhardwaj@windriver.com>