Configure the IP format in ostree remote url
As part of the USM functionality, a script is implemented to
synchronize ostree commits to the subcloud when a prestage is made.

The script uses system command to obtain the IP of the system
controller, which is used to configure the URL of the remote ostree.

In the case of IPv6 in the URL, square brackets is required to
separate the port.

Test Plan:

PASS: Prestage a subcloud supplying release version
	dcmanager subcloud prestage --release MM.mm --for-sw-deploy
	subcloudX
PASS: Prestage a subcloud without release param
	dcmanager subcloud prestage --for-sw-deploy subcloudX

Note: This has been tested with both IPv4 and IPv6 addresses.
      To manually identify the IP of the system controller,
      run the "system addrpool-list" command from the subcloud
      and look for the "system-controller-subnet" parameter.
      Then run the tests. When the prestage process finishes,
      the ostree remote repository will be configured on the
      subcloud, which corresponds to the system controller.
      To verify if the remote is active from the subcloud,
      execute the following command:

      ostree --repo=/var/www/pages/feed/rel-MM.mm/ostree_repo
      remote refs starlingx

      The output should be: starlingx:starlingx

      This should be run only when the prestage process
      completes, as it is during that process that the subcloud
      ostree repo is configured to add the remote.

Story: 2010676
Task: 50634
Change-Id: I5696ee4dfa3e1d6473306d6f98efdf37e0dd1265