Fix address pool deletion on network reconfig
During reconfiguration of network and address pool, the address pools
are not deleted and recreated. The issue is that the network was deleted
before the lookup of the associated address pools. The fix is to
retrieve all the address pools before deleting the network.

Test Plan:
PASS: ansible playbook replay successful
PASS: AIO-SX installation successful

Story: 2011027
Task: 50673

Change-Id: I5a5ccdc66f304613f7da0e92c0c6d4e44a41112f
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>