Subcloud update playbook don't need to remove route
The subcloud routes for system controller are automatically
updated/removed in the sysinv code when moving to/from the admin
network.
The following change is responsible for this improvement:

https://review.opendev.org/c/starlingx/config/+/920669

The unused route is deleted in the function update_subcloud_routes:

https://opendev.org/starlingx/config/src/branch/master/sysinv/sysinv/sysinv/sysinv/common/utils.py#L1858

With the sysinv improvement, when the subcloud update playbook tried
to delete the route, it failed because the route ID was no longer in
the database.
For this reason, the route deletion in the playbook can be removed.

Test done:

DC with AIO-SX subcloud using mgmt network updated to use admin network
DC with AIO-SX subcloud using admin network updated to use mgmt network
DC with AIO-DX subcloud using mgmt network updated to use admin network
DC with AIO-DX subcloud using admin network updated to use mgmt network


Change-Id: I8472444c1f3c7137388a228172745eee7731fa1d
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>