Ignore alarms during subcloud enrollment
During subcloud enrollment, 250.001 alarm is expected to be raised
and persisted after updating the service parameters, networks and
dc-role. In case any failure after these steps, the 250.001 alarm
will block the retry.

This commit ignores this alarm during subcloud enrollment and
rehoming, unblocks the retry of subcloud enrollment. Regarding
the retry of subcloud rehoming, it wil be enabled and verified in a
separete commit.

Test plan:
1. Passed - Deployed a DC(DX system controller + SX subcloud) with
this change.
2. Passed - Verified subcloud enrollment against a non-DC AIO-SX host with
250.001 alarm.
3. Passed - Verfied subcloud rehoming against a AIO-SX subcloud.

Story: 2011100
Task: 50682
Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: Ib4a022ce753cd02f2832eafce97dddf0f02e87da