Require OAM to be dual-stack if cluster-pod is dual-stack
It is observed a failure for the pods (when IPv4 is the secondary
address pool) to resolve 169.254.1.1, the gateway used inside to
access the IPv4 network outside. This is typically when OAM subnet
is single-stack IPv6.

When Pod do ARP request for gateway address, host-side calico
interface, which are proxy-arp enabled, respond to ARP request,
if proxy-arp requirement is met: IP to which it has some route.
In case of single-stack IPv6 OAM, host don't have any route
matching to IP 169.254.1.1, leading this issue.

As Pod to internet IPv4 access will require default IPv4 route
in host anyway, which would be provided by OAM, as solution we
are forcing OAM as dual-stack whenever cluster pod subnet is
dual-stack.

Test Plan:
  PASS: AIO-SX bootstrap with dual-stack cluster pod subnet
        and single-stack OAM subnet throws error (expected)

        external_oam_subnet: 10.20.1.0/24
        external_oam_gateway_address: 10.20.1.1
        external_oam_floating_address: 10.20.1.3
        external_oam_node_0_address: 10.20.1.4
        external_oam_node_1_address: 10.20.1.5
        management_subnet: 192.168.204.0/24,fd01::/64
        management_multicast_subnet: 239.1.1.0/28,ff08::1:1:0/124
        cluster_host_subnet: 192.168.206.0/24,fd02::/64
        cluster_pod_subnet: 172.16.0.0/16,fd03::/64
        cluster_service_subnet: 10.96.0.0/12,fd04::/112
        management_start_address: 192.168.204.1,fd01::1

  PASS: AIO-SX bootstrap with dual-stack cluster pod subnet
        and dual-stack OAM subnet.
        localhost.yml:
        external_oam_subnet: 10.20.1.0/24,fd00::/64
        external_oam_gateway_address: 10.20.1.1,fd00::1
        external_oam_floating_address: 10.20.1.3,fd00::10.20.1.3
        external_oam_node_0_address: 10.20.1.4,fd00::10.20.1.4
        external_oam_node_1_address: 10.20.1.5,fd00::10.20.1.5
        management_subnet: 192.168.204.0/24,fd01::/64
        management_multicast_subnet: 239.1.1.0/28,ff08::1:1:0/124
        cluster_host_subnet: 192.168.206.0/24,fd02::/64
        cluster_pod_subnet: 172.16.0.0/16,fd03::/64
        cluster_service_subnet: 10.96.0.0/12,fd04::/112
        management_start_address: 192.168.204.1,fd01::1

  PASS: AIO-SX bootstrap with single-stack

Story: 2011027
Task: 50728

Change-Id: I2d0bc547a282f6b5cc106b0fe31072ce38ac8abe
Signed-off-by: Tara Nath Subedi <tara.subedi@windriver.com>