Support prestage for minor releases
As part of the subcloud for sw deploy prestage process, in addition
to copying the commits and metadata files of the deployed minor
releases on system controller to the subcloud, it is necessary
to copy the minor releases and software-scripts directory.

This commit adds the logic to be able to copy the directory of
each minor release deployed and the software-scripts to the subcloud.

Test Plan:
PASS: - Deploy a 24.09 DC with at least 1 subcloud.
      - Upload and deploy a 24.09.1 patch release on the system
        controller
      - Prestage the subcloud using command:
        dcmanager subcloud prestage --release 24.09
        --for-sw-deploy subcloudX
      - Verify that 24.09.1 ostree commit is prestaged
        on the subcloud as the patch release has been deployed
        on the system controller
      - Verify that 24.09.1 metadata file is prestaged
        on the subcloud in available state.
      - Verify that rel-24.09.1 directory is prestaged
        on the subcloud
      - Verify that software-scripts directory is prestaged
        on the subcloud
      - Check the software list command output to see the
        release states on subcloud

Story: 2010676
Task: 50770

Change-Id: I497f5df5d29eb5bff28a4b2373bff83b0130ca81