Update the decision logic for subcloud prestage
The subcloud prestage process verifies the patch level deployed
on the system controller and the subcloud.

A comparison is made between both to determine whether prestage
is necessary or not.

The prestage will be necessary as long as the patch level in the
system controller is higher than that of the subcloud.

If the patch level of both is the same or if the subcloud has a
higher patch level than the system controller, the prestage
will not be necessary.

Test Plan:
Setup: - Deploy a 24.09 DC with at least 1 subcloud.

PASS: - Prestage the subcloud using command:
        dcmanager subcloud prestage --release 24.09
        --for-sw-deploy subcloudX
      - Verify that the prestage is skipped because subcloud
        has the same release as the system controller.

PASS: - Upload and deploy a 24.09.1 patch release on the system
        controller
      - Prestage the subcloud using command:
        dcmanager subcloud prestage --release 24.09
        --for-sw-deploy subcloudX
      - Verify that 24.09.1 patch is prestaged on the subcloud

PASS: - Upload and deploy a 24.09.1 patch release on the subcloud
      - Prestage the subcloud using command:
        dcmanager subcloud prestage --release 24.09
        --for-sw-deploy subcloudX
      - Verify that prestage is skipped because the subcloud
        has a higher patch level than the system controller.

Story: 2010676
Task: 50770

Change-Id: Icb324edc259ef9be9434b00ae52e10842edcb135