Improve Ceph disk detection on restore
During the restore procedure, when recovering Ceph data, the playbook
identifies all local disks with Ceph data partitions. There is a file
called ceph_fsid in the OSD disks to identify to which Ceph cluster it
belongs and this value must match the fsid parameter in the
/etc/ceph/ceph.conf configuration file.

The ceph-disk command is used to gather that information. If it runs
with the default /etc/ceph/ceph.conf file that has no fsid configured,
it will always return the disk belonging to the cluster 'ceph'. If the
/etc/ceph/ceph.conf file is restored with the correct fsid, the
ceph-disk command detects the OSD disks correctly.

Not detecting correctly the disks and failing when there is no Ceph
disk available to restore Ceph data will lead to the restore to crash
when restoring the monitor database.

This change will restore the /etc/ceph/ceph.conf file before checking
the OSD disks to detect them properly and fail with a proper failure
message.

Test-Plan:
  PASS: AIO-DX - BnR with correct Ceph disks;
        It should recover without errors.
  PASS: AIO-DX - BnR with wrong Ceph disks;
        It should fail when detecting the Ceph disks
  PASS: AIO-SX - Optimized BnR with correct Ceph disks;
        It should recover without errors.
  PASS: AIO-SX - Optimized BnR with wrong Ceph disks;
        It should fail when detecting the Ceph disks

Closes-bug: 2078004

Change-Id: I6c1e48327077446b14e2c9268fd03b0af10db62f
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>