Avoid showing unnecesary tasks in prestage for software deploy
This commit improves the sequence of the prestage for software
deploy to avoid showing tasks that were executed even when
the prestage was not necessary. For example, when the patch
level of the subcloud is the same or greater than the system
controller.

Now, if the patch level of the subcloud is the same or higher
than the system controller, the playbook will end immediately,
avoiding showing those tasks that are not necessary.

Otherwise, the prestage completes its normal sequence.

Test Plan:
Setup: - Deploy a 24.09 DC with at least 1 subcloud.

PASS: - Prestage the subcloud using command:
        dcmanager subcloud prestage --release 24.09
        --for-sw-deploy subcloudX
      - Verify that the prestage is skipped because subcloud
        has the same release as the system controller.

PASS: - Upload and deploy a 24.09.1 patch release on the system
        controller
      - Prestage the subcloud using command:
        dcmanager subcloud prestage --release 24.09
        --for-sw-deploy subcloudX
      - Verify that 24.09.1 patch is prestaged on the subcloud

PASS: - Upload and deploy a 24.09.2 patch release on the subcloud
      - Prestage the subcloud using command:
        dcmanager subcloud prestage --release 24.09
        --for-sw-deploy subcloudX
      - Verify that prestage is skipped because the subcloud
        has a higher patch level than the system controller.

Story: 2010676
Task: 50770

Change-Id: I83988cb248bb48f57d347f3a27f3d14cf9bfab21
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>