Fix bootstrap not adding secondary IPs to the certificates
The expression used in the platform certificate template file to check
if the secondary IP addresses exist returns false for valid addresses,
causing them to not go into the certificates. This change fixes the
issue by using an expression that returns true for any string that is
not empty.

Test plan

System: AIO-SX dual stack

[PASS] Bootstrap in VirtualBox, check that certificates have both
       primary and secondary addresses

Story: 2011027
Task: 50947
Change-Id: Iad605ef3d931e8b09d927aa8b1a6e5a9dfb8cc34
Signed-off-by: Lucas Ratusznei Fonseca <lucas.ratuszneifonseca@windriver.com>