Fix the failure of syncing software dir in subcloud installation
This is a concurrency issue where two tasks are accessing or modifying
the directory simultaneously. While the install-sync-software-metadata
playbook was running, various host services were started, including the
software-controller-daemon service. This service creates the
"/opt/software/synced" directory with "root:root" permissions if it
does not already exist. This causes a rsync failure because rsync is
running as sysadmin, which does not have write permissions for that
directory.

The fix involves adding a task to create the "/opt/software/synced"
directory with "sysadmin:root" permissions before the rsync task.
This ensures that the directory already exists, allowing
the software-controller-daemon service to detect it and not attempt to
create it with "root:root", thereby allowing rsync to run without
issues.

Test Plan:
PENDING: Deploy more than 50 subclouds simultaneously and ensure that
         this failure does not occur again. Repeat this test multiple
         times.

Closes-Bug: 2078956

Change-Id: Ic8371d05058273daede21d4bf66f1fa83e36c281
Signed-off-by: lzhu1 <li.zhu@windriver.com>