Prestage an N release for software deploy to an N-1 subcloud
This commit adds the logic to be able to prestage an N release
deployed on the system controller, in a subcloud with an N-1
release and with USM capability.

For this scenario to be met, the release of the subcloud must
be different from that of prestage release, it must be an N-1
release supported with USM capability, and the feed directory
pointing to the prestaged N release must not exist on the
subcloud.

Test Plan:

PASS: - Deploy a 24.09 DC with at least 1 22.12 subcloud with
        USM capability.
      - Prestage the subcloud using command:
        dcmanager subcloud prestage --release 24.09
        --for-sw-deploy subcloudX
      - Verify that release is prestaged on the subcloud

PASS: - Upload and deploy a 24.09.1 patch release on the system
        controller.
      - Prestage the subcloud using command:
        dcmanager subcloud prestage --release 24.09
        --for-sw-deploy subcloudX
      - Verify that 24.09.1 patch is prestaged on the subcloud

Regression - Bring up a 24.09 system controller with at least
             one subcloud.
           - Upload and deploy a 24.09.1 patch release on the
             system controller.
           - Prestage the subcloud and verify that release is
             prestaged on the subcloud

Story: 2010676
Task: 51007

Change-Id: Ic36434df277d9236b302e1a140a0370765e517e7
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>